Fix fan_init test (#16865)
* Fix fan_init test

* Readd __init__.py
